fast-api-backend/requirements.txt
pymongo[srv]~=4.11.2
fastapi[standard]~=0.115.11
pydantic~=2.10.6
python-dotenv~=1.0.1